% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drives.R
\name{get_drive}
\alias{get_drive}
\title{Get OneDrive of SharePoint Drive Object}
\usage{
get_drive(file, ...)
}
\arguments{
\item{file}{File Path as a sharepoint_file or onedrive_file}

\item{...}{Reserved for potential future use}
}
\value{
ms_drive object
}
\description{
Both SharePoint and OneDrive have the same methods, so this can be abstracted
down to getting the drive object necessary to interface with the file. This
function uses the file path object to determine which drive object
should be returned
}
\examples{
\dontrun{
sp_file <- sharepoint_file('/some/sharepoint/file.txt')
get_drive(sp_file)
}
}
\concept{Drive Retrievers}
