% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{file_is_current}
\alias{file_is_current}
\title{Check if the local version of a SharePoint or OneDrive file is current}
\usage{
file_is_current(file)
}
\arguments{
\item{file}{OneDrive or SharePoint file path (to be passed to API)}
}
\value{
Boolean
}
\description{
OneDrive files must be downloaded before they can be accessed. This function
takes the OneDrive or SharePoint file path and compares against a file stored
in the temporary directory. If there is no local copy of the file, FALSE
will be returned
}
\examples{
sp_file <- sharepoint_file('/some/sharepoint/file.txt')
file_is_current(sp_file)

}
