% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_local_filename}
\alias{get_local_filename}
\alias{get_cloud_filename}
\title{Get Filepaths for Cloud File or Local Version}
\usage{
get_local_filename(file)

get_cloud_filename(file)
}
\arguments{
\item{file}{SharePoint file path or OneDrive file path}
}
\value{
File name
}
\description{
OneDrive files must be downloaded before they can be accessed. This function
gives the filename back as the local directory path where the file is synced
locally or the path necessary for pulling the file from the cloud
}
\examples{

sp_file <- sharepoint_file('/some/sharepoint/file.txt')
get_local_filename(sp_file)

}
\concept{File Paths}
