% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_file.R
\name{get_file_origin}
\alias{get_file_origin}
\title{Get the origin of a source file}
\usage{
get_file_origin(file)
}
\arguments{
\item{file}{A source_file object or character string}
}
\value{
file origin as a character string
}
\description{
Return the origin of a file path
}
\examples{

fpath <- source_file('some/file/path.txt', origin='onedrive')
get_file_origin(fpath)

}
