% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_file_timestamp}
\alias{get_file_timestamp}
\title{Get the timestamp of a file as a datetime}
\usage{
get_file_timestamp(file, ...)
}
\arguments{
\item{file}{File path}

\item{...}{}

\item{sp_drive}{SharePoint drive API object}
}
\value{
Datetime object
}
\description{
This function gets the timestamp of a file as a datetime object so that the
file timestamps may be compared properly. Different S3 dispatches are
available for handling local files vs. SharePoint
}
\examples{
get_file_timestamp('/some/local/file.txt')

sp_file <- sharepoint_file('/some/sharepoint/file.txt')
get_file_timestamp(sp_file, sp_drive = getOption('atorusdepot.spdrive'))

}
