% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handlers.R
\name{cloud_file}
\alias{cloud_file}
\alias{sharepoint_file}
\alias{onedrive_file}
\title{Cloud File Handlers}
\usage{
cloud_file(path, drive, ...)

sharepoint_file(path, drive = getOption("atorusdepot.spdrive"), ...)

onedrive_file(path, drive = getOption("atorusdepot.onedrive"), ...)
}
\arguments{
\item{path}{File path in SharePoint or OneDrive}

\item{drive}{\code{ms_drive} object from the Microsoft365R package}

\item{...}{Additional parameters - reserved for potential future use}
}
\value{
cloud_file object
}
\description{
These functions allows you to pass a SharePoint or OneDrive file path
directly into the destination file parameter of a function. When writing of
the file is done, the file will automatically be uploaded to SharePoint.
}
\details{
These functions add an extra level of abstraction to a file connection
object. Their behavior should be exactly that of a standard file connection,
such as when you specify a directory path into a writer function like
\code{write.csv}. Some extra metadata has been added to provide the necessary file
handling to upload a file to SharePoint or OneDrive.

When the file is closed, which is typically handled by the writer function
itself, the file connection is closed and the file is uploaded directly into
the cloud using the Microsoft365R \code{ms_drive} object provided (Note this only
happens when the mode of the file was opened to write the file. If the mode
was read only then no uploading is performed). If your setup has been done for
{atorusdepot} properly, the \code{sharepoint_file()} and \code{onedrive_file()}
functions will automatically detect the proper drive object. \code{cloud_file()}
has a drive parameter that can be used to provide your own connection to an
\code{ms_drive} object from the Microsoft365R package.

For information about how the file path should be specified to properly
assign the destination of your file, see \code{?ms365_help}.
}
\examples{

mtcars \%>\%
  write.csv(sharepoint_file("Leadership Team/oversight/test.csv"))

mtcars \%>\%
  write.csv(onedrive_file("Documents/test.csv"))

mtcars \%>\%
  write.csv(
    cloud_file("Documents/test.csv"),
    drive = getOption('atorusdepot.onedrive')
  )
  
}
\concept{Microsoft 365 Cloud Paths}
