% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drives.R
\name{get_sp_object}
\alias{get_sp_object}
\alias{get_od_object}
\title{Set SharePoint or OneDrive drive object to Option}
\usage{
get_sp_object(sp_sitename)

get_od_object()
}
\arguments{
\item{sp_sitename}{SharePoint site name necessary to set up the API object}
}
\description{
This function uses a specified SharePoint site name and uses the
Microsoft365R package to insert the API object into an option. Since this
object will be handed around different functions, the option eases the
hand-off across the package.
}
\concept{Micrsoft API object funcs}
