% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_file.R
\name{source_file}
\alias{source_file}
\title{Create a Source File Indicator}
\usage{
source_file(path, origin = "local")
}
\arguments{
\item{path}{File path respective to the specified origin}

\item{origin}{Origin of the file - either onedrive, sharepoint, or local}
}
\value{
source_file object
}
\description{
This object is basically just a file path, except it additionally has an
attribute for the file origin. This is used to then be able to infer a
connection object based on the \code{source_file} object created.
}
\details{
Since file connections are eliminated and garbage collected once they're
closed, they're unreliable as persistent objects in your R session. To get
around this, {m365filer} creates the file connections on demand. To
preserve the automation of this, this object is created, which feeds into
\code{get_file_connection()} to return a connection object, which is then used to
interface with OneDrive or SharePoint as necessary
}
\examples{

source_file('some/file/path.txt', 'local')
source_file('some/file/path.txt', 'onedrive')
source_file('some/file/path.txt', 'sharepoint')

}
