% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handlers.R
\name{summary.cloud_file}
\alias{summary.cloud_file}
\title{Summarize the Cloud File object}
\usage{
\method{summary}{cloud_file}(con)
}
\arguments{
\item{path}{cloud_file object}
}
\value{
Summary output of the file connection
}
\description{
Abstraction of the summary method for a connection to layer details for a
cloud file connection
}
