% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{update_local_file}
\alias{update_local_file}
\title{Update local copy of a cloud file}
\usage{
update_local_file(file)
}
\arguments{
\item{file}{A cloud_file object}
}
\description{
OneDrive files must be downloaded before they can be accessed. This function
updates the local version of the file by extracting the respective drive
object and downloading the file using the Microsoft 365 API through
Microsoft365R
}
\examples{
sp_file <- sharepoint_file('some/file/path.txt')
update_local_file(sp_file)

}
