% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_file.R
\name{get_file_connection}
\alias{get_file_connection}
\title{Get File Connection}
\usage{
get_file_connection(file)
}
\arguments{
\item{file}{}
}
\value{

}
\description{
Using a source file object, this function returns the file connection object.
{atorusdepot} has file connection objects built for cloud files in the
cloud_file object. This object is abstracted to sharepoint files and onedrive
files, both of which sit on the ms_drive object from the Microsoft365R
package.
}
\details{
These objects are built on top of file connections, which allows you to
interface with files much the same as a local file, by simply providing to
the file parameter of reader or writer functions. Files are read or written
directly to the cloud, and temporary files that clean up with your R sessions
are leveraged to minimize the amount of interface that your local device has
to do with the cloud
}
\examples{

x <- source_file("Documents/test.csv", "onedrive")

mtcars \%>\% write.csv(get_file_connection(x))

y <- read.csv(get_file_connection(x))

}
